Refactor and move over to type script
There was a lot that was changed and its been weeks of working on it and I have fialed to be a good coder and make small commits so im just commiting in one big commit. Big changes include moving to typescirpt, using eslint, bug fixes, and making the userdata base NOT use the User object anymore.